Add files via upload
Discover a magic and mythical City of Mexico. Mexico City is a cosmopolitan city, completed alive. It was created in 1325 as a Tenochtitlan City by the Aztecs. At that moment was like a Venice city. Waters surrounding. Today is a megalopolis with 20 million. This city expresses the past, the present and the future. Its Architecture and cultural diversity makes a unique place